# Add fast path for col like '%%' or col like '%'  or match(col, '.*')