--- a/src/Functions/MatchImpl.h
+++ b/src/Functions/MatchImpl.h
@@ -118,6 +118,18 @@ struct MatchImpl
         if (haystack_offsets.empty())
             return;
 
+        /// Fast path for, because the result is always true or false
+        /// col [i]like '%%'
+        /// col not [i]like '%%'
+        /// col like '%'
+        /// col not [i]like '%'
+        if (is_like && (needle == "%%" or needle == "%"))
+        {
+            for (auto & re : res)
+                re = !negate;
+            return;
+        }
+
         /// Special case that the [I]LIKE expression reduces to finding a substring in a string
         String strstr_pattern;
         if (is_like && impl::likePatternIsSubstring(needle, strstr_pattern))