### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Add fast path for:
- col like '%%'
- col like '%'
- col not like '%'
- col not like '%'
- match(col, '.*')


#### like
Q1: ` select country, count(1) from indigo_show_user_event_62_050101999  where abextra_abflags_v2 like '%%'  group by country;`
Q2: `select country, count(1) from indigo_show_user_event_62_050101999  where not ignore(abextra_abflags_v2)  group by country;`
Notice that indigo_show_user_event_62_050101999  is MergeTree table. 

Before optimization: 
- Q1: `130 rows in set. Elapsed: 3.757 sec. Processed 11.54 million rows, 13.33 GB (3.07 million rows/s., 3.55 GB/s.) `
- Q2: `130 rows in set. Elapsed: 3.064 sec. Processed 11.54 million rows, 13.33 GB (3.77 million rows/s., 4.35 GB/s.)`

After optimization: 
- Q1: `130 rows in set. Elapsed: 2.910 sec. Processed 11.54 million rows, 13.33 GB (3.96 million rows/s., 4.58 GB/s.)`
- Q2: `130 rows in set. Elapsed: 3.011 sec. Processed 11.54 million rows, 13.33 GB (3.83 million rows/s., 4.43 GB/s.)`

Q1 is speeded up by 1.29x in above case. 

#### match 
Q3: `select match(materialize('abcdefg1234567'), '.*') from numbers(100000000) FORMAT Null;`
Q4: `select ignore(materialize('abcdefg1234567')) from numbers(100000000) FORMAT Null;`

Before optimization: 
- Q3: `0 rows in set. Elapsed: 10.271 sec. Processed 100.01 million rows, 800.08 MB (9.74 million rows/s., 77.90 MB/s.)`
- Q4: `0 rows in set. Elapsed: 0.566 sec. Processed 100.01 million rows, 800.08 MB (176.58 million rows/s., 1.41 GB/s.) ` 

After optimization: 
- Q3: `0 rows in set. Elapsed: 0.571 sec. Processed 100.01 million rows, 800.08 MB (175.28 million rows/s., 1.40 GB/s.)`
- Q4: `0 rows in set. Elapsed: 0.566 sec. Processed 100.01 million rows, 800.08 MB (176.58 million rows/s., 1.41 GB/s.) `

Q3 is speeded up by 18x in above case

> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/